Add build scripts for oprofile and its dependencies
Oprofile is a system profiler that people can use to check where the
bottlenecks are in their applications. It need some dependencies,
libpopt (for parsing cmd line options) and some pieces from GNU
binutils (libbfd and the objdump tool).
Note that this is not added to the main build, one has to go to
external-components and explicity call "make oprofile" to build and
install oprofile into the rootfs.
This is done because these profiling tools are not generally needed,
and would require to rebuild the kernel with profiling support anyway
to be able to work, which we don't do normally.